# Optimize Request.SetHeader