--- a/sdk/core/Azure.Core.TestFramework/src/MockRequest.cs
+++ b/sdk/core/Azure.Core.TestFramework/src/MockRequest.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace Azure.Core.TestFramework
 {
@@ -14,8 +13,7 @@ namespace Azure.Core.TestFramework
             ClientRequestId = Guid.NewGuid().ToString();
         }
 
-        private readonly Dictionary<string, List<string>> _headers = new Dictionary<string, List<string>>(StringComparer.OrdinalIgnoreCase);
-
+        private readonly DictionaryHeaders _headers = new();
         public bool IsDisposed { get; private set; }
 
         public override RequestContent Content
@@ -27,64 +25,19 @@ namespace Azure.Core.TestFramework
             }
         }
 
-        protected override void SetHeader(string name, string value)
-        {
-            if (!_headers.TryGetValue(name, out List<string> values))
-            {
-                _headers[name] = _ = new List<string>() { value };
-            }
-            else
-            {
-                values.Clear();
-                values.Add(value);
-            }
-        }
+        protected override void SetHeader(string name, string value) => _headers.SetHeader(name, value);
 
-        protected override void AddHeader(string name, string value)
-        {
-            if (!_headers.TryGetValue(name, out List<string> values))
-            {
-                _headers[name] = values = new List<string>();
-            }
+        protected override void AddHeader(string name, string value) => _headers.AddHeader(name, value);
 
-            values.Add(value);
-        }
+        protected override bool TryGetHeader(string name, out string value) => _headers.TryGetHeader(name, out value);
 
-        protected override bool TryGetHeader(string name, out string value)
-        {
-            if (_headers.TryGetValue(name, out List<string> values))
-            {
-                value = JoinHeaderValue(values);
-                return true;
-            }
+        protected override bool TryGetHeaderValues(string name, out IEnumerable<string> values) => _headers.TryGetHeaderValues(name, out values);
 
-            value = null;
-            return false;
-        }
+        protected override bool ContainsHeader(string name) => _headers.TryGetHeaderValues(name, out _);
 
-        protected override bool TryGetHeaderValues(string name, out IEnumerable<string> values)
-        {
-            var result = _headers.TryGetValue(name, out List<string> valuesList);
-            values = valuesList;
-            return result;
-        }
+        protected override bool RemoveHeader(string name) => _headers.RemoveHeader(name);
 
-        protected override bool ContainsHeader(string name)
-        {
-            return TryGetHeaderValues(name, out _);
-        }
-
-        protected override bool RemoveHeader(string name)
-        {
-            return _headers.Remove(name);
-        }
-
-        protected override IEnumerable<HttpHeader> EnumerateHeaders() => _headers.Select(h => new HttpHeader(h.Key, JoinHeaderValue(h.Value)));
-
-        private static string JoinHeaderValue(IEnumerable<string> values)
-        {
-            return string.Join(",", values);
-        }
+        protected override IEnumerable<HttpHeader> EnumerateHeaders() => _headers.EnumerateHeaders();
 
         public override string ClientRequestId { get; set; }