--- a/sdk/core/Azure.Core/src/Pipeline/HttpWebRequestTransport.cs
+++ b/sdk/core/Azure.Core/src/Pipeline/HttpWebRequestTransport.cs
@@ -310,66 +310,22 @@ namespace Azure.Core.Pipeline
             }
 
             private string? _clientRequestId;
-            private readonly Dictionary<string, List<string>> _headers = new Dictionary<string, List<string>>(StringComparer.OrdinalIgnoreCase);
+            private readonly DictionaryHeaders _headers = new();
 
-            protected internal override void SetHeader(string name, string value)
-            {
-                if (!_headers.TryGetValue(name, out List<string> values))
-                {
-                    _headers[name] = values = new List<string>() { value };
-                }
-                else
-                {
-                    values.Clear();
-                    values.Add(value);
-                }
-            }
+        protected internal override void SetHeader(string name, string value) => _headers.SetHeader(name, value);
 
-            protected internal override void AddHeader(string name, string value)
-            {
-                if (!_headers.TryGetValue(name, out List<string> values))
-                {
-                    _headers[name] = values = new List<string>();
-                }
+        protected internal override void AddHeader(string name, string value) => _headers.AddHeader(name, value);
 
-                values.Add(value);
-            }
+        protected internal override bool TryGetHeader(string name, out string value) => _headers.TryGetHeader(name, out value);
 
-            protected internal override bool TryGetHeader(string name, [NotNullWhen(true)] out string? value)
-            {
-                if (_headers.TryGetValue(name, out List<string> values))
-                {
-                    value = JoinHeaderValue(values);
-                    return true;
-                }
+        protected internal override bool TryGetHeaderValues(string name, out IEnumerable<string> values) => _headers.TryGetHeaderValues(name, out values);
 
-                value = null;
-                return false;
-            }
+        protected internal override bool ContainsHeader(string name) => _headers.TryGetHeaderValues(name, out _);
 
-            protected internal override bool TryGetHeaderValues(string name, out IEnumerable<string> values)
-            {
-                var result = _headers.TryGetValue(name, out List<string> valuesList);
-                values = valuesList;
-                return result;
-            }
+        protected internal override bool RemoveHeader(string name) => _headers.RemoveHeader(name);
 
-            protected internal override bool ContainsHeader(string name)
-            {
-                return TryGetHeaderValues(name, out _);
-            }
+        protected internal override IEnumerable<HttpHeader> EnumerateHeaders() => _headers.EnumerateHeaders();
 
-            protected internal override bool RemoveHeader(string name)
-            {
-                return _headers.Remove(name);
-            }
-
-            protected internal override IEnumerable<HttpHeader> EnumerateHeaders() => _headers.Select(h => new HttpHeader(h.Key, JoinHeaderValue(h.Value)));
-
-            private static string JoinHeaderValue(IEnumerable<string> values)
-            {
-                return string.Join(",", values);
-            }
             public override string ClientRequestId
             {
                 get => _clientRequestId ??= Guid.NewGuid().ToString();