--- a/sdk/core/Azure.Core/src/Request.cs
+++ b/sdk/core/Azure.Core/src/Request.cs
@@ -83,6 +83,7 @@ namespace Azure.Core
             RemoveHeader(name);
             AddHeader(name, value);
         }
+
         /// <summary>
         /// Removes the header from the collection.
         /// </summary>
@@ -103,7 +104,7 @@ namespace Azure.Core
         /// <summary>
         /// Gets the response HTTP headers.
         /// </summary>
-        public RequestHeaders Headers => new RequestHeaders(this);
+        public RequestHeaders Headers => new(this);
 
         /// <summary>
         /// Frees resources held by this <see cref="Response"/> instance.