--- a/sdk/core/Azure.Core/src/Shared/Multipart/MemoryRequest.cs
+++ b/sdk/core/Azure.Core/src/Shared/Multipart/MemoryRequest.cs
@@ -28,6 +28,24 @@ namespace Azure.Core
 #if HAS_INTERNALS_VISIBLE_CORE
         internal
 #endif
+
+       protected override void SetHeader(string name, string value)
+        {
+            if (!_headers.TryGetValue(name, out List<string> values))
+            {
+                _headers[name] = _ = new List<string>() { value };
+            }
+            else
+            {
+                values.Clear();
+                values.Add(value);
+            }
+        }
+
+#if HAS_INTERNALS_VISIBLE_CORE
+        internal
+#endif
+
         protected override void AddHeader(string name, string value)
         {
             if (!_headers.TryGetValue(name, out List<string> values))