… sdk/core/Azure.Core/tests/Azure.Core.Tests.csproj
… sdk/tables/Azure.Data.Tables/src/Azure.Data.Tables.csproj