fixes #19750

Note that for `HttpClientTransport` `Headers.SetValue` for headers other than `Authorization`, the perf is essentially the same. However, in the hot path of setting the `Authorization` header, we see a 2x improvement.

For `HttpWebRequestTransport` (incliding `MockRequest` and `MemoryRequest`) we see a nearly 4x improvement.

```
|                          Method |      Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------------- |----------:|---------:|---------:|------:|--------:|-------:|------:|------:|----------:|
|     HttpClientTsp_SetHeader_Old | 442.31 ns | 4.483 ns | 4.193 ns | 20.62 |    0.20 | 0.0324 |     - |     - |     272 B |
|     HttpClientTsp_SetHeader_New | 452.10 ns | 7.264 ns | 6.795 ns | 21.08 |    0.41 | 0.0324 |     - |     - |     272 B |
| HttpClientTsp_SetAuthHeader_New | 219.25 ns | 2.130 ns | 1.888 ns | 10.22 |    0.14 | 0.0153 |     - |     - |     128 B |
| HttpWebRequestTsp_SetHeader_Old |  83.51 ns | 1.244 ns | 1.164 ns |  3.89 |    0.05 | 0.0105 |     - |     - |      88 B |
| HttpWebRequestTsp_SetHeader_New |  21.45 ns | 0.188 ns | 0.167 ns |  1.00 |    0.00 |      - |     - |     - |         - |
```